+++ src/main/java/br/com/jvsmed/api/repositories/FichaMedicaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FichaMedicaRepository extends JpaRepository<FichaMedicaEntity, Long> {
+public interface FichaMedicaRepository extends JpaRepository<FichaMedicaEntity, Integer> {
     List<FichaMedicaEntity> findByIdPaciente(FichaMedicaEntity idPaciente);
 }
